refactor: extract fs tree builder into util
